+++ org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/commit/RelationChangeItem.java
@@ -20,7 +20,7 @@ public class RelationChangeItem extends ChangeItem {
    private final int relTypeId;
    private final String rationale;
 
-   public RelationChangeItem(long currentSourceGammaId, ModificationType currentSourceModType, long currentSourceTansactionNumber, int aArtId, int bArtId, int relLinkId, int relTypeId, String rationale) {
+   public RelationChangeItem(long currentSourceGammaId, ModificationType currentSourceModType, int currentSourceTansactionNumber, int aArtId, int bArtId, int relLinkId, int relTypeId, String rationale) {
       super(currentSourceGammaId, currentSourceModType, currentSourceTansactionNumber);
 
       this.setItemId(relLinkId);
